+++ source/extensions/lightspeed.trex.source_layer_picker.window/lightspeed/trex/source_layer_picker/window/__init__.py
@@ -15,14 +15,12 @@
 * limitations under the License.
 """
 
-from .unit.test_combobox_filter_tooltips import TestComboboxFilterTooltipsUnit
-from .unit.test_geometry_prims_filter import TestGeometryPrimsFilterUnit
-from .unit.test_is_capture_filter import TestIsCaptureFilterUnit
-from .e2e.test_scene_edit_state_filter import TestSceneEditFilterE2E
-
 __all__ = [
-    "TestComboboxFilterTooltipsUnit",
-    "TestGeometryPrimsFilterUnit",
-    "TestIsCaptureFilterUnit",
-    "TestSceneEditFilterE2E",
+    "SearchableLayerModel",
+    "SourceLayerPicker",
+    "SourceLayerPickerDelegate",
 ]
+
+from .delegate import SourceLayerPickerDelegate
+from .model import SearchableLayerModel
+from .source_layer_picker import SourceLayerPicker