+++ source/extensions/lightspeed.trex.stage_manager.plugin.interaction.usd/lightspeed/trex/stage_manager/plugin/interaction/usd/all_categories.py
@@ -49,6 +49,7 @@ class RemixAllCategoriesInteractionPlugin(RemixStageManagerUSDInteractionPlugin)
             "LightPrimsFilterPlugin",
             "MeshPrimsFilterPlugin",
             "OmniPrimsFilterPlugin",
+            "SceneEditFilterPlugin",
             "SearchFilterPlugin",
             "SkeletonPrimsFilterPlugin",
         ],