+++ source/extensions/lightspeed.trex.stage_manager.plugin.interaction.usd/lightspeed/trex/stage_manager/plugin/interaction/usd/all_materials.py
@@ -36,6 +36,7 @@ class RemixAllMaterialsInteractionPlugin(AllMaterialsInteractionPlugin, RemixSta
             *AllMaterialsInteractionPlugin.model_fields["compatible_filters"].default,
             # Remix filters
             "IsCaptureFilterPlugin",
+            "SceneEditFilterPlugin",
         ],
         exclude=True,
     )