+++ source/extensions/lightspeed.trex.stage_manager.plugin.interaction.usd/lightspeed/trex/stage_manager/plugin/interaction/usd/all_prims.py
@@ -40,6 +40,7 @@ class RemixAllPrimsInteractionPlugin(AllPrimsInteractionPlugin, RemixStageManage
             "MeshGroupFilterPlugin",
             "ParticleSystemsFilterPlugin",
             "RemixLogicPrimsFilterPlugin",
+            "SceneEditFilterPlugin",
         ],
         exclude=True,
     )