+++ source/extensions/lightspeed.trex.stage_manager.plugin.interaction.usd/lightspeed/trex/stage_manager/plugin/interaction/usd/all_tags.py
@@ -36,6 +36,7 @@ class RemixAllTagsInteractionPlugin(AllTagsInteractionPlugin, RemixStageManagerU
             # Remix filters
             "IsCaptureFilterPlugin",
             "ParticleSystemsFilterPlugin",
+            "SceneEditFilterPlugin",
         ],
         exclude=True,
     )